+++ test/include/test_static.h
@@ -15,7 +15,7 @@ extern "C" {
 char *crcAdd(char *json, uint16_t seqno);
 bool crcError(char *json, uint16_t shouldBeSeqno);
 void delayIO(void);
-const char * i2cNoteQueryLength(uint32_t * available, size_t timeoutMs);
+const char * i2cNoteQueryLength(uint32_t * available, uint32_t timeoutMs);
 void setTime(JTIME seconds);
 bool timerExpiredSecs(uint32_t *timer, uint32_t periodSecs);
 char Jtolower(char c);
